pyproject.toml
[project]
name = "speech-to-speech-translator"
version = "0.1.0"
description = "A Python tool to translate audio files while preserving voice and intonation using Gemini 2.5 Flash."
authors = [
    {name = "Artem Ryazanov", email = "artryazanov@gmail.com"},
]
dependencies = [
    "google-generativeai>=0.8.0",
    "pydub>=0.25.1",
    "python-dotenv>=1.0.0",
    "typer>=0.9.0",
    "rich>=13.0.0",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"

[tool.pdm]
distribution = true
